Extract `aggregateBase.scala`
